feat(model): vectorized training (batch and mini-batch gradient descent)
Previously, the update of parameters was done after backpropagating each batch, as opposed to SGD (Stochastic Gradient Descent).
However, batches elements were fed one at a time to the network as the implementation was not fully vectorized.
Moreover, the division of delta weights and biases by batch size was skipped.